Working procedures of image insertion and receival, and update and deletion
The database can now store and send images to the view and the deletion and update procedures are functional. The necessary stored procedures in the database for these procedures were also implemented. When updating room info, however, the Rate parameter needs to be revised: When shown on View, it shows its value in double, but when altered and sent, it requires a value in integer.